Moved the pendant to its own module.
If the application requires the pendant implementation it can simply add this module as a dependency.
The PendantConfigBean is no longer configurable from the settings file (will figure this out later).
The SystemStateBean is now populated by the PendantUI and no longer in the backend.
