+++ ugs-core/test/com/willwinder/universalgcodesender/utils/SettingsTest.java
@@ -57,7 +57,6 @@ public class SettingsTest {
         assertNotNull(target.getMainWindowSettings());
         assertFalse(target.isManualModeEnabled());
         assertEquals(Double.valueOf(1.0), Double.valueOf(target.getManualModeStepSize()));
-        assertNotNull(target.getPendantConfig());
         assertEquals("", target.getPort());
         assertEquals("115200", target.getPortRate());
         assertTrue(target.isScrollWindowEnabled());
