+++ ugs-pendant/src/main/java/com/willwinder/universalgcodesender/pendantui/PendantConfigBean.java
