+++ ugs-pendant/src/main/resources/pendantUI/css/smoothness/jquery-ui-1.10.4.custom.css
