+++ ugs-pendant/src/main/resources/pendantUI/index.html
