+++ ugs-pendant/src/main/resources/pendantUI/js/jquery-1.10.2.js
